split custom (generic) flag types to separate files
