+++ util/flagutil/duration.go
@@ -5,22 +5,7 @@ import (
 	"fmt"
 	"strings"
 	"time"
-
-	"github.com/dustin/go-humanize"
 )
-
-type ByteSizeValue uint64
-
-func (b *ByteSizeValue) Set(s string) error {
-	v, err := humanize.ParseBytes(s)
-	if err != nil {
-		return err
-	}
-	*b = ByteSizeValue(v)
-	return nil
-}
-
-func (b *ByteSizeValue) String() string { return humanize.Bytes(uint64(*b)) }
 
 // Duration implements cli.GenericFlag methods to support time.Duration with days, e.g. 1d
 type Duration time.Duration
